+++ services/121-service/src/activities/enum/activity-type.enum.ts
@@ -1,0 +1,8 @@
+export enum ActivityTypeEnum {
+  DataChange = 'data-change',
+  FinancialServiceProviderChange = 'financial-service-provider-change',
+  Message = 'message',
+  Note = 'note',
+  StatusChange = 'status-change',
+  Transaction = 'transaction',
+}
