+++ services/121-service/src/activities/interfaces/base-activity.interface.ts
@@ -1,0 +1,12 @@
+import { ActivityTypeEnum } from '@121-service/src/activities/enum/activity-type.enum';
+
+export interface BaseActivity {
+  id: string;
+  user: {
+    id?: number;
+    username?: string;
+  };
+  created: Date;
+  type: ActivityTypeEnum;
+  attributes: Record<string, unknown>;
+}
