+++ services/121-service/src/activities/interfaces/status-change-activity.interface.ts
@@ -1,0 +1,11 @@
+import { ActivityTypeEnum } from '@121-service/src/activities/enum/activity-type.enum';
+import { BaseActivity } from '@121-service/src/activities/interfaces/base-activity.interface';
+import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
+
+export interface StatusChangeActivity extends BaseActivity {
+  type: ActivityTypeEnum.StatusChange;
+  attributes: {
+    oldValue: RegistrationStatusEnum;
+    newValue: RegistrationStatusEnum;
+  };
+}
