+++ services/121-service/src/activities/types/activity.type.ts
@@ -1,0 +1,14 @@
+import { DataChangeActivity } from '@121-service/src/activities/interfaces/data-change-activity.interface';
+import { FinancialServiceProviderChangeActivity } from '@121-service/src/activities/interfaces/financial-service-provider.interface';
+import { MessageActivity } from '@121-service/src/activities/interfaces/message-activity.interface';
+import { NoteActivity } from '@121-service/src/activities/interfaces/note-activity.interface';
+import { StatusChangeActivity } from '@121-service/src/activities/interfaces/status-change-activity.interface';
+import { TransactionActivity } from '@121-service/src/activities/interfaces/transaction-activity.interface';
+
+export type Activity =
+  | DataChangeActivity
+  | FinancialServiceProviderChangeActivity
+  | MessageActivity
+  | NoteActivity
+  | StatusChangeActivity
+  | TransactionActivity;
